requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
SQLAlchemy
alembic
python-multipart
orjson
asyncpg
pydantic-settings
python-dotenvasyncpg==0.30.0
